CORE-7943 Rename ledger-common-base-integrationtest to ledger-common-base-integration-test and split everything accordingly
