+++ libs/ledger/ledger-common-data/build.gradle
@@ -35,7 +35,7 @@ dependencies {
     testImplementation "org.mockito:mockito-core:$mockitoVersion"
     testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
 
-    integrationTestImplementation project(':testing:ledger:ledger-common-base-integrationtest')
+    integrationTestImplementation project(':testing:ledger:ledger-common-base-integration-test')
     integrationTestRuntimeOnly project(':libs:flows:session-manager-impl')
     integrationTestRuntimeOnly project(':libs:lifecycle:lifecycle-impl')
     integrationTestRuntimeOnly project(':libs:messaging:db-message-bus-impl')
