+++ testing/ledger/ledger-common-base-integration-test/src/main/java/net/corda/ledger/common/integration/test/package-info.java
@@ -1,4 +1,4 @@
 @Export
-package net.corda.ledger.common.integrationtest;
+package net.corda.ledger.common.integration.test;
 
 import org.osgi.annotation.bundle.Export;
