+++ testing/ledger/ledger-consensual-testkit/build.gradle
@@ -15,5 +15,5 @@ dependencies {
     implementation project(':libs:ledger:ledger-consensual-data')
     api project(':components:ledger:ledger-consensual-flow')
     implementation project(':testing:ledger:ledger-common-testkit')
-    api project(':testing:ledger:ledger-common-base-integrationtest')
+    api project(':testing:ledger:ledger-common-base-integration-test')
 }
